% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{all.numeric}
\alias{all.numeric}
\title{Check/Convert Values to Numeric}
\usage{
\method{all}{numeric}(x, what = c("test", "vector"), extras = c(".", "NA"))
}
\arguments{
\item{x}{Character vector to check/convert.}

\item{what}{"test" (check if numeric) or "vector" (convert to numeric).}

\item{extras}{Additional values to be treated as missing or empty.}
}
\description{
Check/Convert Values to Numeric
}
