% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirFunUtils.R
\name{CalculateMirTargetSet}
\alias{CalculateMirTargetSet}
\title{FUNCTION_TITLE}
\usage{
CalculateMirTargetSet(nms, operation)
}
\arguments{
\item{nms}{PARAM_DESCRIPTION}

\item{operation}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
