% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{checkMiRNAVersion}
\alias{checkMiRNAVersion}
\title{FUNCTION_TITLE}
\usage{
checkMiRNAVersion(miRNANames, verbose = TRUE)
}
\arguments{
\item{miRNANames}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
