% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{ClearFactorStrings}
\alias{ClearFactorStrings}
\title{FUNCTION_TITLE}
\usage{
ClearFactorStrings(cls.nm, query)
}
\arguments{
\item{cls.nm}{PARAM_DESCRIPTION}

\item{query}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
