% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{ComputeColorGradient}
\alias{ComputeColorGradient}
\title{FUNCTION_TITLE}
\usage{
ComputeColorGradient(nd.vec, background = "black")
}
\arguments{
\item{nd.vec}{PARAM_DESCRIPTION}

\item{background}{PARAM_DESCRIPTION, Default: 'black'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
