% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{CreateMirNets}
\alias{CreateMirNets}
\title{FUNCTION_TITLE}
\usage{
CreateMirNets(net.type)
}
\arguments{
\item{net.type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
