% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{DecomposeMirGraph}
\alias{DecomposeMirGraph}
\title{FUNCTION_TITLE}
\usage{
DecomposeMirGraph(net.type, gObj, minNodeNum = 2)
}
\arguments{
\item{net.type}{PARAM_DESCRIPTION}

\item{gObj}{PARAM_DESCRIPTION}

\item{minNodeNum}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
