% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnotUtils.R
\name{doProbeMapping}
\alias{doProbeMapping}
\title{FUNCTION_TITLE}
\usage{
doProbeMapping(probe.vec, platform)
}
\arguments{
\item{probe.vec}{PARAM_DESCRIPTION}

\item{platform}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
