% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnotUtils.R
\name{doSymbol2EntrezMapping}
\alias{doSymbol2EntrezMapping}
\title{FUNCTION_TITLE}
\usage{
doSymbol2EntrezMapping(entrez.vec)
}
\arguments{
\item{entrez.vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
