% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{ExtractMirNetModule}
\alias{ExtractMirNetModule}
\title{FUNCTION_TITLE}
\usage{
ExtractMirNetModule(nodeids)
}
\arguments{
\item{nodeids}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
