% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{FilterMirNet}
\alias{FilterMirNet}
\title{FUNCTION_TITLE}
\usage{
FilterMirNet(net.type, nd.type, min.dgr, min.btw)
}
\arguments{
\item{net.type}{PARAM_DESCRIPTION}

\item{nd.type}{PARAM_DESCRIPTION}

\item{min.dgr}{PARAM_DESCRIPTION}

\item{min.btw}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
