% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{FilterMirNetByList}
\alias{FilterMirNetByList}
\title{FUNCTION_TITLE}
\usage{
FilterMirNetByList(net.type, ids, id.type)
}
\arguments{
\item{net.type}{PARAM_DESCRIPTION}

\item{ids}{PARAM_DESCRIPTION}

\item{id.type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
