% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{FindCommunities}
\alias{FindCommunities}
\title{Find Communities}
\usage{
FindCommunities(method = "walktrap", use.weight = FALSE)
}
\arguments{
\item{method}{Method: walktrap, infomap, labelprop.}

\item{use.weight}{Logical.}
}
\description{
Find Communities
}
