% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{getColorGradient}
\alias{getColorGradient}
\title{FUNCTION_TITLE}
\usage{
getColorGradient(n, col.from = "white", col.to = "red")
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{col.from}{PARAM_DESCRIPTION, Default: 'white'}

\item{col.to}{PARAM_DESCRIPTION, Default: 'red'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
