% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{GetColorSchema}
\alias{GetColorSchema}
\title{FUNCTION_TITLE}
\usage{
GetColorSchema(my.grps)
}
\arguments{
\item{my.grps}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
