% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{getExpColors}
\alias{getExpColors}
\title{FUNCTION_TITLE}
\usage{
getExpColors(nd.vec, col.vec)
}
\arguments{
\item{nd.vec}{PARAM_DESCRIPTION}

\item{col.vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
