% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{GetExtendRange}
\alias{GetExtendRange}
\title{FUNCTION_TITLE}
\usage{
GetExtendRange(vec, unit = 10)
}
\arguments{
\item{vec}{PARAM_DESCRIPTION}

\item{unit}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
