% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{GetNetNms}
\alias{GetNetNms}
\title{FUNCTION_TITLE}
\usage{
GetNetNms()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
