% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{GetNodeNumByType}
\alias{GetNodeNumByType}
\title{FUNCTION_TITLE}
\usage{
GetNodeNumByType(g)
}
\arguments{
\item{g}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
