% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mir_target_enrich.R
\name{GetRandomMirTargetGenes}
\alias{GetRandomMirTargetGenes}
\title{Get Random miRNA Target Genes}
\usage{
GetRandomMirTargetGenes(qSize, perm.num)
}
\description{
Get Random miRNA Target Genes
}
