% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{GetSeqnm}
\alias{GetSeqnm}
\title{FUNCTION_TITLE}
\usage{
GetSeqnm(rowid)
}
\arguments{
\item{rowid}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
