% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtils.R
\name{GetSigGenes}
\alias{GetSigGenes}
\title{FUNCTION_TITLE}
\usage{
GetSigGenes(p.lvl, fc.lvl, direction, update = T)
}
\arguments{
\item{p.lvl}{PARAM_DESCRIPTION}

\item{fc.lvl}{PARAM_DESCRIPTION}

\item{direction}{PARAM_DESCRIPTION}

\item{update}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
