% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{GetTtestP}
\alias{GetTtestP}
\title{Get p-value from t-test}
\usage{
GetTtestP(
  my.dat,
  my.cls,
  grp1,
  grp2,
  paired = FALSE,
  equal.var = TRUE,
  nonpar = F
)
}
\description{
Get p-value from t-test
}
