% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{GetUniqueEntries}
\alias{GetUniqueEntries}
\title{Get Unique Entries}
\usage{
GetUniqueEntries(db.path, statement)
}
\arguments{
\item{db.path}{Path to database.}

\item{statement}{Statement.}
}
\description{
Get Unique Entries
}
