% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{GetUniqueEpigeneNames}
\alias{GetUniqueEpigeneNames}
\title{FUNCTION_TITLE}
\usage{
GetUniqueEpigeneNames(orgType = "hsa")
}
\arguments{
\item{orgType}{PARAM_DESCRIPTION, Default: 'hsa'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
