% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{GetUniqueMoleculeNames}
\alias{GetUniqueMoleculeNames}
\title{Get Unique Molecule Names}
\usage{
GetUniqueMoleculeNames(orgType = "hsa")
}
\arguments{
\item{orgType}{Organism type.}
}
\description{
Get Unique Molecule Names
}
