% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{GetUniqueSourceNames}
\alias{GetUniqueSourceNames}
\title{FUNCTION_TITLE}
\usage{
GetUniqueSourceNames(orgType)
}
\arguments{
\item{orgType}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
