% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{GetUpdateClassNames}
\alias{GetUpdateClassNames}
\title{FUNCTION_TITLE}
\usage{
GetUpdateClassNames(orgType, source)
}
\arguments{
\item{orgType}{PARAM_DESCRIPTION}

\item{source}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
