% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{hex2rgba}
\alias{hex2rgba}
\title{FUNCTION_TITLE}
\usage{
hex2rgba(cols)
}
\arguments{
\item{cols}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
