% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{Init.Data}
\alias{Init.Data}
\title{FUNCTION_TITLE}
\usage{
Init.Data(dataType, analType)
}
\arguments{
\item{dataType}{PARAM_DESCRIPTION}

\item{analType}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
