% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirFunUtils.R
\name{LoadGOLib}
\alias{LoadGOLib}
\title{FUNCTION_TITLE}
\usage{
LoadGOLib(onto)
}
\arguments{
\item{onto}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
