% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirFunUtils.R
\name{LoadHMDDLib}
\alias{LoadHMDDLib}
\title{FUNCTION_TITLE}
\usage{
LoadHMDDLib()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
