% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirFunUtils.R
\name{LoadTissueLib}
\alias{LoadTissueLib}
\title{FUNCTION_TITLE}
\usage{
LoadTissueLib()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
