% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{LogNorm}
\alias{LogNorm}
\title{FUNCTION_TITLE}
\usage{
LogNorm(x, min.val)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{min.val}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
