% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{miRNA_PrecursorToMature}
\alias{miRNA_PrecursorToMature}
\title{FUNCTION_TITLE}
\usage{
miRNA_PrecursorToMature(miRNANames, version = NULL)
}
\arguments{
\item{miRNANames}{PARAM_DESCRIPTION}

\item{version}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
