% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{myave}
\alias{myave}
\title{FUNCTION_TITLE}
\usage{
myave(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
