% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirFunUtils.R
\name{PerformAPIMirTargetEnrichAnalysis}
\alias{PerformAPIMirTargetEnrichAnalysis}
\title{Perform miRNA Target Enrichment Analysis (API)}
\usage{
PerformAPIMirTargetEnrichAnalysis(
  adjust.type,
  fun.type,
  file.nm,
  IDs,
  algo,
  mode = "parallel"
)
}
\description{
Perform miRNA Target Enrichment Analysis (API)
}
