% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{PerformArrayDataNormalization}
\alias{PerformArrayDataNormalization}
\title{Peform array data normalization}
\usage{
PerformArrayDataNormalization(norm.opt)
}
\description{
Peform array data normalization
}
