% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{PerformCountDataNormalization}
\alias{PerformCountDataNormalization}
\title{Perform count data normalization
#' @export}
\usage{
PerformCountDataNormalization(norm.opt, disp.opt)
}
\description{
Perform count data normalization
#' @export
}
