% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnotUtils.R
\name{PerformDisMapping}
\alias{PerformDisMapping}
\title{FUNCTION_TITLE}
\usage{
PerformDisMapping()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
