% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtils.R
\name{PerformHTqPCR}
\alias{PerformHTqPCR}
\title{FUNCTION_TITLE}
\usage{
PerformHTqPCR(target.grp, method)
}
\arguments{
\item{target.grp}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
