% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{PerformLimma}
\alias{PerformLimma}
\title{FUNCTION_TITLE}
\usage{
PerformLimma(target.grp)
}
\arguments{
\item{target.grp}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
