% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnotUtils.R
\name{PerformMir2EpiMapping}
\alias{PerformMir2EpiMapping}
\title{FUNCTION_TITLE}
\usage{
PerformMir2EpiMapping()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
