% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnotUtils.R
\name{PerformMirGeneMapping}
\alias{PerformMirGeneMapping}
\title{PerformMirGeneMapping}
\usage{
PerformMirGeneMapping()
}
\value{
"mirnet_mir_target.csv" results table will be downloaded in your working directory
}
\description{
Map miRNAs to target genes or from target genes to miRNAs
}
