% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirFunUtils.R
\name{PerformMirTargetEnrichAnalysis}
\alias{PerformMirTargetEnrichAnalysis}
\title{Perform miRNA Target Enrichment Analysis}
\usage{
PerformMirTargetEnrichAnalysis(
  adjust.type,
  fun.type,
  file.nm,
  IDs,
  algo,
  mode = "serial"
)
}
\description{
Perform miRNA Target Enrichment Analysis
}
