% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{PerformQpcrDataNormalization}
\alias{PerformQpcrDataNormalization}
\title{Perform qPCR data normalization}
\usage{
PerformQpcrDataNormalization(norm.opt = "quantile")
}
\description{
Perform qPCR data normalization
}
