% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{PerformSpeciesMapping}
\alias{PerformSpeciesMapping}
\title{FUNCTION_TITLE}
\usage{
PerformSpeciesMapping(status)
}
\arguments{
\item{status}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
