% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{PlotBetweennessHistogram}
\alias{PlotBetweennessHistogram}
\title{Plot Betweenness Histogram}
\usage{
PlotBetweennessHistogram(imgNm, netNm = "NA", dpi = 72, format = "png")
}
\description{
Plot Betweenness Histogram
}
