% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{PrepareMirNet}
\alias{PrepareMirNet}
\title{FUNCTION_TITLE}
\usage{
PrepareMirNet(mir.nm, file.nm)
}
\arguments{
\item{mir.nm}{PARAM_DESCRIPTION}

\item{file.nm}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
