% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{Query.miRNetDB}
\alias{Query.miRNetDB}
\title{Query miRNet DB}
\usage{
Query.miRNetDB(db.path, q.vec, table.nm, col.nm, db.nm = "mirtarbase")
}
\arguments{
\item{db.path}{Path to database.}

\item{q.vec}{Vector to query.}

\item{table.nm}{Table name.}

\item{col.nm}{Column names.}

\item{db.nm}{Name of the database.}
}
\description{
Query miRNet DB
}
