% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnotUtils.R
\name{queryGeneDB}
\alias{queryGeneDB}
\title{FUNCTION_TITLE}
\usage{
queryGeneDB(table.nm, data.org)
}
\arguments{
\item{table.nm}{PARAM_DESCRIPTION}

\item{data.org}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
