% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiUtils.R
\name{QueryMultiListMir}
\alias{QueryMultiListMir}
\title{FUNCTION_TITLE}
\usage{
QueryMultiListMir()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
