% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{QueryPpiSQLiteZero}
\alias{QueryPpiSQLiteZero}
\title{Query Protein-Protein Interaction}
\usage{
QueryPpiSQLiteZero(table.nm, q.vec, requireExp, min.score)
}
\description{
Query Protein-Protein Interaction
}
