% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{QueryTFSQLite}
\alias{QueryTFSQLite}
\title{Query Transcription Factor}
\usage{
QueryTFSQLite(table.nm, q.vec, col.nm)
}
\arguments{
\item{table.nm}{Table name.}

\item{q.vec}{Vector to query.}

\item{col.nm}{Column names.}
}
\description{
Query Transcription Factor
}
