% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{QueryXenoMirSQLite}
\alias{QueryXenoMirSQLite}
\title{FUNCTION_TITLE}
\usage{
QueryXenoMirSQLite(db.path, q.vec, table.nm, col.nm, source)
}
\arguments{
\item{db.path}{PARAM_DESCRIPTION}

\item{q.vec}{PARAM_DESCRIPTION}

\item{table.nm}{PARAM_DESCRIPTION}

\item{col.nm}{PARAM_DESCRIPTION}

\item{source}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
