% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{ReadTabExpressData}
\alias{ReadTabExpressData}
\title{FUNCTION_TITLE}
\usage{
ReadTabExpressData(dataName)
}
\arguments{
\item{dataName}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
