% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{RecordRCommand}
\alias{RecordRCommand}
\title{Record R Commands}
\usage{
RecordRCommand(cmd)
}
\arguments{
\item{cmd}{Commands}

\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Record R Commands
}
