% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{RemoveDuplicates}
\alias{RemoveDuplicates}
\title{FUNCTION_TITLE}
\usage{
RemoveDuplicates(data, lvlOpt, quiet = T)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{lvlOpt}{PARAM_DESCRIPTION}

\item{quiet}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
