% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{rescale2NewRange}
\alias{rescale2NewRange}
\title{Rescale to New Range}
\usage{
rescale2NewRange(qvec, a, b)
}
\arguments{
\item{qvec}{Vector to be rescaled.}

\item{a}{Lower bound.}

\item{b}{Upper bound.}
}
\description{
Rescale to New Range
}
