% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{RowScale}
\alias{RowScale}
\title{FUNCTION_TITLE}
\usage{
RowScale(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
