% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SearchMultiNet}
\alias{SearchMultiNet}
\title{Search Multi-Network Data}
\usage{
SearchMultiNet(input.type)
}
\arguments{
\item{input.type}{miRNA and target, e.g., mir2gene, mir2molecule, mir2disease.}
}
\description{
Search Multi-Network Data
}
