% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{SetCurrentDataMulti}
\alias{SetCurrentDataMulti}
\title{FUNCTION_TITLE}
\usage{
SetCurrentDataMulti()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
