% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SetPpiDb}
\alias{SetPpiDb}
\title{Set Protein-Protein Interaction Db}
\usage{
SetPpiDb(db, req, conf)
}
\arguments{
\item{db}{Database name.}

\item{req}{Logical.}

\item{conf}{Min score.}
}
\description{
Set Protein-Protein Interaction Db
}
