% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{SetupFilterList}
\alias{SetupFilterList}
\title{FUNCTION_TITLE}
\usage{
SetupFilterList(exp, miranda, tarpmir)
}
\arguments{
\item{exp}{PARAM_DESCRIPTION}

\item{miranda}{PARAM_DESCRIPTION}

\item{tarpmir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
