% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SetupIndListData}
\alias{SetupIndListData}
\title{Setup List Data}
\usage{
SetupIndListData(listInput, orgType, inputType, idType, tissue, target)
}
\arguments{
\item{listInput}{List to query.}

\item{orgType}{Organism type.}

\item{inputType}{Input type.}

\item{idType}{ID type.}

\item{tissue}{Tissue type.}

\item{target}{Target.}
}
\value{
List data initialized.
}
\description{
Setup List Data
}
