% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{SetupItemFromList}
\alias{SetupItemFromList}
\title{FUNCTION_TITLE}
\usage{
SetupItemFromList(orgType, species)
}
\arguments{
\item{orgType}{PARAM_DESCRIPTION}

\item{species}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
