% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SetupItemFromPickList}
\alias{SetupItemFromPickList}
\title{Setup Item From Picklist}
\usage{
SetupItemFromPickList(orgType = "hsa", tissue, idType)
}
\arguments{
\item{orgType}{Organism type.}

\item{tissue}{Tissue type.}

\item{idType}{ID type.}
}
\description{
Setup Item From Picklist
}
