% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{SetupMirExpressData}
\alias{SetupMirExpressData}
\title{FUNCTION_TITLE}
\usage{
SetupMirExpressData()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
