% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SetupMirListData}
\alias{SetupMirListData}
\title{Setup miRNA List Data}
\usage{
SetupMirListData(mirs, orgType, idType, tissue, targetOpt = NULL)
}
\arguments{
\item{mirs}{A list of miRNAs.}

\item{orgType}{Organism type.}

\item{idType}{miRNA ID type.}

\item{tissue}{Tissue type.}

\item{targetOpt}{Target options (optional).}
}
\value{
miRNA list data initialized.
}
\description{
Setup miRNA List Data
}
