% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{SetupSourceFromList}
\alias{SetupSourceFromList}
\title{FUNCTION_TITLE}
\usage{
SetupSourceFromList(source)
}
\arguments{
\item{source}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
