% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{SetupXenoMirListData}
\alias{SetupXenoMirListData}
\title{FUNCTION_TITLE}
\usage{
SetupXenoMirListData(mirs, orgType, idType)
}
\arguments{
\item{mirs}{PARAM_DESCRIPTION}

\item{orgType}{PARAM_DESCRIPTION}

\item{idType}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
