% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{sync2vecs}
\alias{sync2vecs}
\title{FUNCTION_TITLE}
\usage{
sync2vecs(src.vec, tgt.vec)
}
\arguments{
\item{src.vec}{PARAM_DESCRIPTION}

\item{tgt.vec}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
