% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{all.numeric}
\alias{all.numeric}
\title{FUNCTION_TITLE}
\usage{
\method{all}{numeric}(x, what = c("test", "vector"), extras = c(".", "NA"))
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{what}{PARAM_DESCRIPTION, Default: c("test", "vector")}

\item{extras}{PARAM_DESCRIPTION, Default: c(".", "NA")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
