% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{CalculatePairwiseDiff}
\alias{CalculatePairwiseDiff}
\title{FUNCTION_TITLE}
\usage{
CalculatePairwiseDiff(mat)
}
\arguments{
\item{mat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
