% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{ClearFactorStrings}
\alias{ClearFactorStrings}
\title{Clear Factor Strings}
\usage{
ClearFactorStrings(cls.nm, query)
}
\arguments{
\item{cls.nm}{Character string to be prefixed.}

\item{query}{Character vector of string to be cleaned.}
}
\description{
Clear Factor Strings
}
