% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{ComputeColorGradient}
\alias{ComputeColorGradient}
\title{Compute Color Gradient}
\usage{
ComputeColorGradient(nd.vec, background = "black", centered)
}
\arguments{
\item{nd.vec}{Numeric vector.}

\item{background}{Background color.}

\item{centered}{Logical.}
}
\description{
Compute Color Gradient
}
