% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{GetColorGradient}
\alias{GetColorGradient}
\title{Get Color Gradient}
\usage{
GetColorGradient(background, center)
}
\arguments{
\item{center}{Logical}

\item{backgroun}{Background.}
}
\description{
Get Color Gradient
}
