% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{GetColorGradient}
\alias{GetColorGradient}
\title{FUNCTION_TITLE}
\usage{
GetColorGradient(background, center)
}
\arguments{
\item{background}{PARAM_DESCRIPTION}

\item{center}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
