% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{getExpColors}
\alias{getExpColors}
\title{Colors Based on Expression}
\usage{
getExpColors(nd.vec, col.vec)
}
\arguments{
\item{nd.vec}{Numeric vector containing values to be mapped to colors.}

\item{col.vec}{Color vector.
export}
}
\description{
Colors Based on Expression
}
