% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{GetExtendRange}
\alias{GetExtendRange}
\title{Get Extend Range}
\usage{
GetExtendRange(vec, unit = 10)
}
\arguments{
\item{vec}{Vector input.}

\item{unit}{Unit to extend the range.}
}
\description{
Get Extend Range
}
