% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirFunUtils.R
\name{GetRandomMirTargetGenes}
\alias{GetRandomMirTargetGenes}
\title{FUNCTION_TITLE}
\usage{
GetRandomMirTargetGenes(qSize, perm.num)
}
\arguments{
\item{qSize}{PARAM_DESCRIPTION}

\item{perm.num}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
