% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mir_target_enrich.R
\name{GetRandomXenoMirTargetGenes}
\alias{GetRandomXenoMirTargetGenes}
\title{Get Random Xeno-miRNA Target Genes}
\usage{
GetRandomXenoMirTargetGenes(qSize, perm.num)
}
\description{
Get Random Xeno-miRNA Target Genes
}
