% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsUtils.R
\name{GetTtestP}
\alias{GetTtestP}
\title{FUNCTION_TITLE}
\usage{
GetTtestP(
  my.dat,
  my.cls,
  grp1,
  grp2,
  paired = FALSE,
  equal.var = TRUE,
  nonpar = F
)
}
\arguments{
\item{my.dat}{PARAM_DESCRIPTION}

\item{my.cls}{PARAM_DESCRIPTION}

\item{grp1}{PARAM_DESCRIPTION}

\item{grp2}{PARAM_DESCRIPTION}

\item{paired}{PARAM_DESCRIPTION, Default: FALSE}

\item{equal.var}{PARAM_DESCRIPTION, Default: TRUE}

\item{nonpar}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
