% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{GetUniqueEntries}
\alias{GetUniqueEntries}
\title{FUNCTION_TITLE}
\usage{
GetUniqueEntries(db.path, statement)
}
\arguments{
\item{db.path}{PARAM_DESCRIPTION}

\item{statement}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
