% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{GetUniqueEpigeneNames}
\alias{GetUniqueEpigeneNames}
\title{Get Unique Epigene Names}
\usage{
GetUniqueEpigeneNames(orgType = "hsa")
}
\arguments{
\item{orgType}{Organism type.}
}
\description{
Get Unique Epigene Names
}
