% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{miRNA_PrecursorToMature}
\alias{miRNA_PrecursorToMature}
\title{miRNA Name Conversion From Precursor to Mature}
\usage{
miRNA_PrecursorToMature(miRNANames, version = NULL)
}
\description{
miRNA Name Conversion From Precursor to Mature
}
