% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirNetUtils.R
\name{PerformLayOut}
\alias{PerformLayOut}
\title{Apply Graph Layout}
\usage{
PerformLayOut(g, layers, algo, focus = "")
}
\arguments{
\item{g}{Graph object.}

\item{algo}{Layout algorithm: Default, FrR, circle, random, lgl, gopt, circular_tripartite, tripartite, concentric, backbone, mds.}

\item{focus}{Focus node.}

\item{layer}{Optional: layer information.}
}
\description{
Apply Graph Layout
}
