% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MirFunUtils.R
\name{PerformMirTargetEnrichAnalysis}
\alias{PerformMirTargetEnrichAnalysis}
\title{FUNCTION_TITLE}
\usage{
PerformMirTargetEnrichAnalysis(
  adjust.type,
  fun.type,
  file.nm,
  IDs,
  algo,
  mode = "serial"
)
}
\arguments{
\item{adjust.type}{PARAM_DESCRIPTION}

\item{fun.type}{PARAM_DESCRIPTION}

\item{file.nm}{PARAM_DESCRIPTION}

\item{IDs}{PARAM_DESCRIPTION}

\item{algo}{PARAM_DESCRIPTION}

\item{mode}{PARAM_DESCRIPTION, Default: 'serial'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
