% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprUtils.R
\name{PerformQpcrDataNormalization}
\alias{PerformQpcrDataNormalization}
\title{FUNCTION_TITLE}
\usage{
PerformQpcrDataNormalization(norm.opt = "quantile")
}
\arguments{
\item{norm.opt}{PARAM_DESCRIPTION, Default: 'quantile'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
