% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{QueryTFSQLite}
\alias{QueryTFSQLite}
\title{FUNCTION_TITLE}
\usage{
QueryTFSQLite(table.nm, q.vec, col.nm)
}
\arguments{
\item{table.nm}{PARAM_DESCRIPTION}

\item{q.vec}{PARAM_DESCRIPTION}

\item{col.nm}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
