% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SetupIndListData}
\alias{SetupIndListData}
\title{FUNCTION_TITLE}
\usage{
SetupIndListData(listInput, orgType, inputType, idType, tissue, target)
}
\arguments{
\item{listInput}{PARAM_DESCRIPTION}

\item{orgType}{PARAM_DESCRIPTION}

\item{inputType}{PARAM_DESCRIPTION}

\item{idType}{PARAM_DESCRIPTION}

\item{tissue}{PARAM_DESCRIPTION}

\item{target}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
