% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtils.R
\name{SetupItemFromPickList}
\alias{SetupItemFromPickList}
\title{FUNCTION_TITLE}
\usage{
SetupItemFromPickList(orgType = "hsa", tissue, idType)
}
\arguments{
\item{orgType}{PARAM_DESCRIPTION, Default: 'hsa'}

\item{tissue}{PARAM_DESCRIPTION}

\item{idType}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
