% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtils.R
\name{SetupMirListData}
\alias{SetupMirListData}
\title{FUNCTION_TITLE}
\usage{
SetupMirListData(mirs, orgType, idType, tissue = "na", targetOpt = NULL)
}
\arguments{
\item{mirs}{PARAM_DESCRIPTION}

\item{orgType}{PARAM_DESCRIPTION}

\item{idType}{PARAM_DESCRIPTION}

\item{tissue}{PARAM_DESCRIPTION}

\item{targetOpt}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
