% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoMirUtils.R
\name{SetupXenoMirListData}
\alias{SetupXenoMirListData}
\title{Setup Xeno miRNA List}
\usage{
SetupXenoMirListData(mirs, orgType, idType)
}
\arguments{
\item{mirs}{miRNAs.}

\item{orgType}{Organism type.}

\item{idType}{ID type.}
}
\description{
Setup Xeno miRNA List
}
