% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscUtils.R
\name{sync2vecs}
\alias{sync2vecs}
\title{Synchronize Two Vectors}
\usage{
sync2vecs(src.vec, tgt.vec)
}
\arguments{
\item{src.vec}{Source character vector to be arranged.}

\item{tgt.vec}{Target numeric vector of the same length.}
}
\description{
Synchronize Two Vectors
}
